% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max_perm_per_core.R
\name{benchmark_nm}
\alias{benchmark_nm}
\title{Benchmark the amount of RAM needed per permutation in a null model}
\usage{
benchmark_nm(phenotype, covariates, benchmark_size = 1000)
}
\arguments{
\item{phenotype}{A phenotype file read into R from standard PLINK phenotype
format (see (\href{https://www.cog-genomics.org/plink2/formats}{PLINK
documentation}), with labeled columns for FID, IID, and trait.}

\item{covariates}{A covariate file read into R from either standard PLINK
covariate format ((\href{https://www.cog-genomics.org/plink2/formats}{PLINK
documentation}) or in a `.csv` with headers for genotype ID (in the first
column) and an ID for each covariate (subsequent columns)}

\item{benchmark_size}{The number of permutations to be used for the benchmark}
}
\value{
Integer, the amount of RAM needed per permutation in a null model (in
  bytes)
}
\description{
Benchmark the amount of RAM needed per permutation in a null model
}
\examples{
data("small_phenodata")
data("small_covariates")

nm_RAM_per_perm <- benchmark_nm(phenotype = small_phenodata,
                                covariates = small_covariates,
                                benchmark_size = 1000)

}
