% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max_perm_per_core.R
\name{calculate_max_perm_per_core}
\alias{calculate_max_perm_per_core}
\title{Calculate the number of permutations in a null model that can fit in RAM
partitioned for a single thread}
\usage{
calculate_max_perm_per_core(nm_RAM_per_perm, RAM, n_thread)
}
\arguments{
\item{nm_RAM_per_perm}{Integer, the amount of RAM needed per permutation in a
null model (in bytes); this is the output from \code{\link{benchmark_nm}}}

\item{RAM}{Integer, the total amount of RAM available to be used, for all
threads (in bytes)}

\item{n_thread}{Integer, the number of threads over which RAM will be divided}
}
\value{
An integer indicating the amount of RAM that is available to be used
  by any given thread
}
\description{
Calculate the number of permutations in a null model that can fit in RAM
partitioned for a single thread
}
\examples{
data("sample_phenotype")
data("sample_covariates")

nm_RAM_per_perm <- mtmcskat::benchmark_nm(phenotype = sample_phenotype,
                                          covariates = sample_covariates,
                                          benchmark_size = 1000)

RAM <- benchmarkme::get_ram()

calculate_max_perm_per_core(nm_RAM_per_perm = nm_RAM_per_perm,
                            RAM = RAM,
                            n_thread = 2)

}
