% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SKAT_empirical_p.R
\name{calculate_SKAT_empirical_p}
\alias{calculate_SKAT_empirical_p}
\title{Produce permuted p-values or empirical p-value for SKAT}
\usage{
calculate_SKAT_empirical_p(
  Z,
  n_permutations,
  null_model,
  return_all_p_vals = FALSE,
  ...
)
}
\arguments{
\item{Z}{Matrix with row for each genotype in the SNP dataset, columns for
each SNP in the SNP window of interest, and numeric values of 0, 1, 2
indicating number of alternative alleles for each SNP in each genotype (or
NA for indels or missing data)}

\item{n_permutations}{Integer indicating the number of permutations used to
calculate empirical p-values}

\item{null_model}{Object generated by \code{\link[SKAT]{SKAT_Null_Model}}}

\item{return_all_p_vals}{If `TRUE`, will return a vector of all permuted
p-values (useful for when not all permutations can fit into memory
available for a single thread); if `FALSE`, will calculate and return the
empirical p-value}

\item{...}{Additional parameters passed on to \code{\link[SKAT]{SKAT}}}
}
\value{
An empirical p-value or vector of permuted p-values, depending on the
  user-submitted argument for `return_all_p` as described above
}
\description{
Produce permuted p-values or empirical p-value for SKAT
}
\examples{
data("small_pre_allocated_windows")

sample_null_model <- SKAT::SKAT_Null_Model(
  small_phenodata ~ 1 + as.matrix(small_covariates), out_type="C",
  n.Resampling = 1000)

calculate_SKAT_empirical_p(
  Z = small_pre_allocated_windows[[1]][[2]],
  n_permutations = 1000,
  null_model = sample_null_model,
  return_all_p = FALSE)

calculate_SKAT_empirical_p(
  Z = small_pre_allocated_windows[[1]][[2]],
  n_permutations = 1000,
  null_model = sample_null_model,
  return_all_p = TRUE)
}
