% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_windows.R
\name{chunk_windows}
\alias{chunk_windows}
\title{Divide a list of SNP windows into a chunk for each thread}
\usage{
chunk_windows(pre_allocated_SNP_windows, n_thread)
}
\arguments{
\item{pre_allocated_SNP_windows}{Output from \code{\link{pre_allocate}},
which is a list of lists, with each sub-list containing elements as
described in documentation for \code{\link{extract_window}}}

\item{n_thread}{An integer indicating the number of threads to be used for
multithreading}
}
\value{
a list of lists of lists, in which the list of lists passed by the
  `pre_allocated_SNP_windows` argument is divided into a number of lists
  indicated by the `n_threads` argument
}
\description{
Divide a list of SNP windows into a chunk for each thread
}
\examples{
data("small_pre_allocated_windows")
pre_allocated_SNP_windows <- chunk_windows(
  pre_allocated_SNP_windows = small_pre_allocated_windows,
  n_thread = 2)

}
