% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_utils.R
\name{determine_n_permutations}
\alias{determine_n_permutations}
\title{Estimate, based on initial p-values, the number of permut}
\usage{
determine_n_permutations(leading_0s, desired_sig_figs)
}
\arguments{
\item{leading_0s}{An integer for the number of 0's between the decimal point
and the first nonzero value in the maximum p-value for SNP windows that will
be included in the subset of interest}

\item{desired_sig_figs}{An integer for the number of significant figures
the user desires for empirical p-values}
}
\value{
An integer, representing the number of permutations expected to be
needed to calculate an empirical p-value with the desired level of accuracy
}
\description{
This function estimates, based on the number of leading 0's in an initial
p-value, how many permutations will be needed to calculate an empirical
p-value out to the user-specified number of significant figures
}
\examples{
# 10,000 permutations should be needed to calculate p-values below 0.01
# (e.g. 0.0095 has two leading 0's) out to two significant figures
determine_n_permutations(leading_0s = 2, desired_sig_figs = 2)
}
