% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_window.R
\name{extract_window}
\alias{extract_window}
\title{Extract and format a window of SNPs to be tested in a SKAT kernel}
\usage{
extract_window(
  this_position,
  window_size,
  genodata,
  impute_to_mean = TRUE,
  remove_novar_SNPs = TRUE,
  missing_cutoff = 0.15
)
}
\arguments{
\item{this_position}{An integer, indicating the center of a SNP window for
which the user wishes to extract SNPs (in base pairs)}

\item{window_size}{An integer, indicating the size of each SNP window (in
base pairs)}

\item{genodata}{A matrix obtained by reading data from `.traw` format (see
(see \url{https://www.cog-genomics.org/plink2/formats}{PLINK
documentation})) into R}

\item{impute_to_mean}{If `TRUE`, NA values for each SNP are replaced with
the mean alternative allele count for the given SNP}

\item{remove_novar_SNPs}{If `TRUE`, SNPs with no variation will be removed}
}
\value{
A list containing three objects: an integer of chromosome of origin,
  an integer of SNP window center position, and a matrix of alternative
  allele counts for each genotype and SNP within the given window
}
\description{
Extract and format a window of SNPs to be tested in a SKAT kernel
}
\examples{
small_genodata_path <- system.file("extdata",
                                   "poplar_200genotypes_14490to14520kb.traw",
                                   package = "mtmcskat")
small_genodata <- data.table::fread(small_genodata_path)

extract_window(this_position = 145e5,
  window_size = 3000,
  genodata = small_genodata)

}
