% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_p_null.R
\name{p_empirical_from_tally}
\alias{p_empirical_from_tally}
\title{Tally p-values from permuted null model to enable calculation of empirical
p-values}
\usage{
p_empirical_from_tally(p_null_tallies, scaffold_ID)
}
\arguments{
\item{p_null_tallies}{a matrix containing tallies of null p-values above or
below the initial p-value}

\item{scaffold_ID}{a string or numeric value indicating the scaffold, to be
used only for labeling output}
}
\value{

}
\description{
Tally p-values from permuted null model to enable calculation of empirical
p-values
}
\examples{
data("sample_p_null_tallies")
p_empirical_from_tally(p_null_tallies = sample_p_null_tallies,
  scaffold_ID = 10)
}
