% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_allocate_windows.R
\name{re_allocate_windows}
\alias{re_allocate_windows}
\title{Obtain a new list of SNP windows, based on p-values from an earlier list}
\usage{
re_allocate_windows(data, upper_bound, lower_bound, pre_allocated_SNP_windows)
}
\arguments{
\item{data}{An output from \code{\link{mtskat}} or a similarly formatted
data.frame}

\item{upper_bound}{A numeric value corresponding to the maximum p-value for
the range of interest}

\item{lower_bound}{A numeric value correspondign to the minimum p-value for
the range of interest}

\item{pre_allocated_SNP_windows}{Output from \code{\link{pre_allocate}},
which is a list of lists, with each sub-list containing elements as
described in documentation for \code{\link{extract_window}}}
}
\value{
Output is the same format as for \code{\link{pre_allocate}}: a list
  of lists, with each sub-list containing elements as described in
  documentation for \code{\link{extract_window}}
}
\description{
Obtain a new list of SNP windows, based on p-values from an earlier list
}
\examples{
data("sample_pre_allocated_SNP_windows")

re_allocate_windows(
  data = sample_mtskat_results,
  upper_bound = 0.01,
  lower_bound = 0.001,
  pre_allocated_SNP_windows = sample_pre_allocated_SNP_windows)

}
