% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_genodata}
\alias{sample_genodata}
\title{SNP data in `.traw` format for poplar samples in stem regeneration experiment}
\format{
A data frame with 5076 rows and 888 variables

\describe{
  \item{CHR}{Integer or character value indicating the chromosome or scaffold
  on which a given SNP is found}
  \item{SNP}{Character providing a label for a given SNP; this is optional
  and the column contains "." for all values by default if SNPs are not
  named. This variable is not used by the `mtmcskat` package.}
  \item{X.C.M}{Position of a given SNP in morgans or centimorgans; this is
  optional and can be filled with "0" if not used. This variable is not used
  by the `mtmcskat` package.}
  \item{POS}{Integer providing the position of a given SNP, in base pairs}
  \item{COUNTED}{Character, either "A", "T", "C" or "G" indicating the
  common (or most common) allele at the position of the given SNP}
  \item{COUNTER}{Character, either "A", "T", "C" or "G" indicating the
  alternative allele, also known as the rare allele,
  at the position of the given SNP. If multiple alternative alleles exist
  for a position, they are provided on separate rows with the same common
  allele and position.}
  \item{X201782_400194}{Columns 7-888 contain alternative allele
  counts for each of 882 genotypes in the poplar GWAS population. These are
  integer values ranging from 0-2, indicating the number of alleles which are
  the alternative allele for a given SNP in the column for a given genotype.
  The name of each column from 7-888 follows a format that includes the
  Family ID and Individual ID for each genotype, following the format
  X<FID>_<IID>}
}
}
\usage{
sample_genodata
}
\description{
This dataset contains alternative allele counts for 5076 base positions from
about 14.46Mb 14.55Mb to on Chromosome 10 of Populus trichocarpa. This data
was obtained first by downloading
(\href{https://cbi.ornl.gov/data}{poplar SNP data}) from the Department
of Energy's Center for Bioenergy Innovation, converting the data from Variant
Call Format (VCF) to `.traw` using PLINK (see
\href{https://www.cog-genomics.org/plink2/formats}{PLINK documentation})
and finally loading the `.traw` file into R using
\code{\link[data.table]{fread}}. Each row represents a single nucleotide
polymorphism (SNP) with variation across the poplar GWAS population.
}
\details{
Support for the Poplar GWAS dataset is provided by the U.S.
Department of Energy, Office of Science Biological and Environmental
Research (BER) via the Bioenergy Science Center (BESC) under Contract
No. DE-PS02-06ER64304.  The Poplar GWAS Project used resources of the
Oak Ridge Leadership Computing Facility and the Compute and Data Environment
for Science at Oak Ridge National Laboratory, which is supported by the
Office of Science of the U.S. Department of Energy under
Contract No. DE-AC05-00OR22725
}
\keyword{datasets}
