% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mtmcskat_results}
\alias{sample_mtmcskat_results}
\title{MTMCSKAT results from multithreading over SNP windows}
\format{
A data frame with 3 rows and 4 variables
\describe{
  \item{Chr}{Integer or character value indicating the chromosome or scaffold
  on which a given SNP is found}
  \item{position}{Integer providing the position of a given SNP, in base
  pairs}
  \item{SKAT_p-val}{Numeric value (double) providing the p-value obtained by
  SKAT without resampling}
  \item{SKAT_p-val_resampled}{Numeric value (double) providing the empirical
  p-value obtained by MTMCSKAT after resampling}
}
}
\usage{
sample_mtmcskat_results
}
\description{
Results obtained from running MTMCSKAT with parallelization over SNP windows
in \code{\link{sample_re_allocated_SNP_windows}}, using the function
\code{\link{mtmcskat_SNPs}}
}
\keyword{datasets}
