% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mtmcskatnm1_results}
\alias{sample_mtmcskatnm1_results}
\title{MTMCSKAT results from multithreading over null models in a single batch}
\format{
A data frame with 3 rows and 4 variables
\describe{
  \item{Chr}{Integer or character value indicating the chromosome or scaffold
  on which a given SNP is found}
  \item{position}{Integer providing the position of a given SNP, in base
  pairs}
  \item{SKAT_p-val}{Numeric value (double) providing the p-value obtained by
  SKAT without resampling}
  \item{SKAT_p-val_resampled}{Numeric value (double) providing the empirical
  p-value obtained by MTMCSKAT after resampling}
}
}
\usage{
sample_mtmcskatnm1_results
}
\description{
Results obtained from running MTMCSKAT over
\code{\link{sample_re_allocated_SNP_windows}} with multithreading over
null models using the function \code{\link{mtmcskat_NullModels}},
while all null models can fit into the allotted RAM simultaneously and
thus all resampling can be completed in a single batch
}
\keyword{datasets}
