% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_p_null_tallies}
\alias{sample_p_null_tallies}
\title{Tallies of p-values from resampling}
\format{
A data frame with 192 rows and 4 variables
\describe{
  \item{position}{Integer indicating, in base pairs, the center of a SNP
  window}
  \item{SKAT_p-val}{Numeric value (double) indicating the original p-value
  for the given SNP window, obtained by SKAT without resampling}
  \item{n_perm_above}{The number of permuted p-values from a given thread
  that are above the original p-value for a given window}
  \item{n_perm_ltoreq}{The number of permuted p-values from a given thread
  that are below the original p-value for a given window}}
}
\usage{
sample_p_null_tallies
}
\description{
This dataset features tallies of permuted null model p-values that are above
or below the model p-value that was obtained from a null model without
resampling. The sums of p-values above or below this model p-value can be
used to calculate an empirical p-value. There is a row for each SNP window,
thread combination, since SNP windows are tested on all threads using
different null models on each thread. This sample dataset was obtained from
\code{\link{mtmcskat_NullModels}}, which multithreads Monte Carlo SKAT
over null models.
}
\keyword{datasets}
