% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_phenotype}
\alias{sample_phenotype}
\title{Proportions of poplar image pixels labeled as shoot}
\format{
A list with 882 values
}
\usage{
sample_phenotype
}
\description{
For all samples in the 882-genotype poplar GWAS population, this list
contains either 1) NA, or samples not studied in the sample GWAS, or 2)
a value ranging from 0 to 1, indicating the proportion of image pixels that
are labeled as shoot (rather than as callus or unregenerated stem). Samples
were treated by submerging poplar cuttings in water, applying TDZ to their
tips, and leaving a microcentrifuge tube ontop to provide humid conditions.
Phenotyping was performed five weeks after this treatment, via imaging plants
and classifying pixels with a trained convolutional neural network.
}
\keyword{datasets}
