% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_pre_allocated_SNP_windows}
\alias{sample_pre_allocated_SNP_windows}
\title{List of consecutive SNP windows with positon labels}
\format{
A list with 90 values
}
\usage{
sample_pre_allocated_SNP_windows
}
\description{
This dataset contains a list of SNP windows, each formatted as described
in documentation for \code{\link{sample_SNP_window}}. All of these SNP
windows are consecutive and span the range of 14.46Mb 14.55Mb on Chr. 10.
Each SNP window spans a range of 3kb and consecutive windows are shifted by
1kb relative to the previous adjacent window.
}
\keyword{datasets}
