% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_re_allocated_SNP_windows}
\alias{sample_re_allocated_SNP_windows}
\title{List of selected SNP windows significantly associated with shoot trait}
\format{
A list with 4 values
}
\usage{
sample_re_allocated_SNP_windows
}
\description{
This dataset contains a list of SNP windows, each formatted as described
in documentation for \code{\link{sample_SNP_window}}. This list of SNP
windows includes all those within the range of 14.46Mb 14.55Mb on Chr. 10
that produced p-values between 0.01 and 0.001 in \code{\link{mtskat}}.
This list was produced by using
\code{\link{re_allocate_windows}} to yield a subset of SNP windows that
produced p-values within the desired range.
}
\keyword{datasets}
