% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_SNP_window_processed}
\alias{sample_SNP_window_processed}
\title{Labeled window of alt. allele counts (filtered and imputed) for P trichocarpa}
\format{
A list containing three values

\describe{
  \item{Position}{Integer, the position of the center of a SNP window,
  in base pairs}
  \item{Z}{Matrix of integers containing alternative allele counts for
  each of 175 SNPs (columns) in the SNP window, across 200 genotypes (rows)}
  \item{Chr}{Integer or character indicating the chromosome or other
  scaffold, respectively, on which the SNP window is found}
}
}
\source{
\url{https://cbi.ornl.gov/data}
}
\usage{
sample_SNP_window_processed
}
\description{
A dataset containing the position of a SNP window in the genome and all of
the alternative allele counts for each SNP within the SNP window
}
\details{
Unlike \code{\link{sample_SNP_window}}, this object HAS
undergone processing to filter out SNPs with high
missing rates (> 0.15) or to impute missing values to the mean.
Thus, the use of this object will NOT produce warning messages from `SKAT`,
as there is no need for ad-hoc missing rate filtering and imputation
to be performed.

Support for the Poplar GWAS dataset is provided by the U.S.
Department of Energy, Office of Science Biological and Environmental
Research (BER) via the Bioenergy Science Center (BESC) under Contract
No. DE-PS02-06ER64304.  The Poplar GWAS Project used resources of the
Oak Ridge Leadership Computing Facility and the Compute and Data Environment
for Science at Oak Ridge National Laboratory, which is supported by the
Office of Science of the U.S. Department of Energy under
Contract No. DE-AC05-00OR22725
}
\keyword{datasets}
