% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_window_list}
\alias{sample_window_list}
\title{List of SNP window centers}
\format{
A list with 90 values
}
\usage{
sample_window_list
}
\description{
This dataset is a list identical to that produced by the command
`seq(14460000, 14549000, by=1000)`. This is a list of SNP window positions,
where each value represents the center (in base pairs) of a SNP window.
}
\keyword{datasets}
