% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_windows_range_p.R
\name{select_windows_range_p}
\alias{select_windows_range_p}
\title{Select windows corresponding to p-values within range of interest}
\usage{
select_windows_range_p(data, upper_bound, lower_bound)
}
\arguments{
\item{data}{An output from \code{\link{mtskat}} or a similarly formatted
data.frame}

\item{upper_bound}{A numeric value corresponding to the maximum p-value for
the range of interest}

\item{lower_bound}{A numeric value correspondign to the minimum p-value for
the range of interest}
}
\value{
A numeric vector with indices for the position (center of window) for
  each SNP window with p-values within the range of interest
}
\description{
Select windows corresponding to p-values within range of interest
}
\examples{
select_windows_range_p(
  data = sample_mtskat_results,
  upper_bound = 0.01,
  lower_bound = 0.001)

}
