% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKAT_one_window.R
\name{SKAT_one_window}
\alias{SKAT_one_window}
\title{Run SKAT, optionally obtaining permuted p-values or empirical p-value, and
organize results into a vector to be added to dataframe of results}
\usage{
SKAT_one_window(
  this_position,
  Z,
  scaffold_ID,
  null_model,
  n_permutations,
  resampling = FALSE,
  return_all_p_vals = FALSE
)
}
\arguments{
\item{this_position}{Integer indicating the center of a given SNP window (in
base pairs)}

\item{Z}{Matrix with row for each genotype in the SNP dataset, columns for
each SNP in the SNP window of interest, and numeric values of 0, 1, 2
indicating number of alternative alleles for each SNP in each genotype (or
NA for indels or missing data)}

\item{scaffold_ID}{Integer indicating the chromosome or scaffold of interest}

\item{null_model}{Object generated by \code{\link[SKAT]{SKAT_Null_Model}}}

\item{n_permutations}{Integer indicating the number of permutations used to
calculate empirical p-values}

\item{resampling}{If TRUE, will call \code{\link{calculate_SKAT_empirical_p}}
to calculate an empirical p-value or obtain a vector of permuted p-values.
If FALSE, will pass NA in place of an empirical p-value}

\item{return_all_p_vals}{If `TRUE`, will return a vector of all permuted
p-values (useful for when not all permutations can fit into memory
available for a single thread); if `FALSE`, will calculate and return the
empirical p-value}
}
\value{
If not returning all permuted p-values (as indicated by
  `return_all_p_vals`, the output will be a vector of length 4, containing
  `scaffold_ID`, `this_position`, the model p-value, and finally the
  empirical p-value (if applicable; otherwise NA). If returning all p-values,
  the length of the output vector will be two (for SNP scaffold and window
  center position) plus the number of permutations
}
\description{
Run SKAT, optionally obtaining permuted p-values or empirical p-value, and
organize results into a vector to be added to dataframe of results
}
\examples{
data("small_pre_allocated_windows")
sample_null_model <- SKAT::SKAT_Null_Model(
  small_phenodata ~ 1 + as.matrix(small_covariates), out_type="C",
  n.Resampling = 1000)

SKAT_one_window(
  this_position = small_pre_allocated_windows[[1]][[1]],
  Z = small_pre_allocated_windows[[1]][[2]],
  scaffold_ID = small_pre_allocated_windows[[1]][[3]],
  n_permutations = 1000,
  null_model = sample_null_model,
  resampling = TRUE,
  return_all_p = FALSE)
}
