% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_covariates}
\alias{small_covariates}
\title{Covariates for poplar samples in stem regeneration experiment}
\format{
A data frame with 200 rows and 11 variables

\describe{
  \item{Stem_diameter}{Numeric value (double) representing the diameter of
  the stem of a poplar sample, in mm}
  \item{Phase1}{Indicator variable, with 1 if the sample was studied
  in Phase 1}
  \item{Phase2}{Indicator variable, with 1 if the sample was studied
  in Phase 2}
  \item{Phase3}{Indicator variable, with 1 if the sample was studied
  in Phase 3}
  \item{Phase4}{Indicator variable, with 1 if the sample was studied
  in Phase 4}
  \item{Phase5}{Indicator variable, with 1 if the sample was studied
  in Phase 5}
  \item{Phase6}{Indicator variable, with 1 if the sample was studied
  in Phase 6}
  \item{Phase7}{Indicator variable, with 1 if the sample was studied
  in Phase 7}
  \item{PC1}{Loadings for the first principal component derived from SNPs
  using the `pca` method in PLINK}
  \item{PC2}{Loadings for the second principal component derived from SNPs
  using the `pca` method in PLINK}
  \item{PC3}{Loadings for the third principal component derived from SNPs
  using the `pca` method in PLINK}
}
}
\usage{
small_covariates
}
\description{
A dataset containing covariates including stem diameter, phase and principal
components (PCs), for a small subset of 200 genotypes out of the poplar
GWAS population
}
\keyword{datasets}
