% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_mtskat_results}
\alias{small_mtskat_results}
\title{MTSKAT results from multithreading over SNP windows}
\format{
A data frame with 29 rows and 4 variables
\describe{
  \item{Chr}{Integer or character value indicating the chromosome or scaffold
  on which a given SNP is found}
  \item{position}{Integer providing the position of a given SNP, in base
  pairs}
  \item{SKAT_p-val}{Numeric value (double) providing the p-value obtained by
  SKAT without resampling}
  \item{SKAT_p-val_resampled}{Column containing NA, to be replaced with
  empirical p-values when this data structure is provided to
  \code{\link{mtmcskat_NullModels}} or \code{\link{mtmcskat_SNPs}}}
}
}
\usage{
small_mtskat_results
}
\description{
Results obtained from running MTSKAT with multithreading over SNP windows
in \code{\link{sample_re_allocated_SNP_windows}}, using the function
\code{\link{mtskat}}
}
\keyword{datasets}
