% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_phenodata}
\alias{small_phenodata}
\title{Proportions of poplar image pixels labeled as shoot}
\format{
A list with 200 values
}
\usage{
small_phenodata
}
\description{
For 200 out of samples in the 882-genotype poplar GWAS population, this list
contains a value from 0 to 1, indicating the proportion of image pixels that
are labeled as shoot (rather than as callus or unregenerated stem). Samples
were treated by submerging poplar cuttings in water, applying TDZ to their
tips, and leaving a microcentrifuge tube ontop to provide humid conditions.
Phenotyping was performed five weeks after this treatment, via imaging plants
and classifying pixels with a trained convolutional neural network.
}
\keyword{datasets}
