% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_pre_allocated_windows_processed}
\alias{small_pre_allocated_windows_processed}
\title{List of consecutive SNP windows (filtered and imputed) with positon labels}
\format{
A list with 90 values
}
\usage{
small_pre_allocated_windows_processed
}
\description{
Similarly to \code{\link{small_pre_allocated_windows}},
this dataset contains a list of SNP windows, each formatted as described
in documentation for \code{\link{sample_SNP_window}}. All of these SNP
windows are consecutive and span the range of 14.49Mb 14.52Mb on Chr. 10.
Each SNP window spans a range of 3kb and consecutive windows are shifted by
1kb relative to the previous adjacent window.
}
\details{
Unlike \code{\link{small_pre_allocated_windows}}, this object HAS
undergone processing to filter out SNPs with high
missing rates (> 0.15) or to impute missing values to the mean.
Thus, the use of this object will NOT produce warning messages from `SKAT`,
as there is no need for ad-hoc missing rate filtering and imputation
to be performed.
}
\keyword{datasets}
