% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_list_of_lists.R
\name{subset_list_of_lists}
\alias{subset_list_of_lists}
\title{Subset a list of lists to a list of desired lists only Subset a list of lists
down to a list of lists containing only lists with the value at a certain
position (with this values indicated in another list)}
\usage{
subset_list_of_lists(list_of_lists, desired_list, subindex = 1)
}
\arguments{
\item{list_of_lists}{a list of lists, such as a list of lists containing SNP
window position, SNPs and scaffold, as generated by
\code{\link{pre_allocate}}}

\item{desired_list}{list of desired elements}

\item{subindex}{index of lists within list of lists, which will be checked to
see if the list contains the desired element and should be subsetted}
}
\value{
A list in which all of the values indicated by the `subindex`
  parameter are matched by the `desired_list`
}
\description{
Subset a list of lists to a list of desired lists only Subset a list of lists
down to a list of lists containing only lists with the value at a certain
position (with this values indicated in another list)
}
\examples{
subset_list_of_lists(
  list_of_lists = sample_pre_allocated_SNP_windows,
  desired_list = c(14502000, 14503000, 14504000, 14505000),
  subindex = 1)

}
