% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_window.R
\name{extract_window}
\alias{extract_window}
\title{Extract and format a window of SNPs to be tested in a SKAT kernel}
\usage{
extract_window(this_position, window_size, genodata)
}
\arguments{
\item{this_position}{An integer, indicating the center of a SNP window for
which the user wishes to extract SNPs (in base pairs)}

\item{window_size}{An integer, indicating the size of each SNP window (in
base pairs)}

\item{genodata}{A matrix obtained by reading data from `.traw` format (see
(see \url{https://www.cog-genomics.org/plink2/formats}{PLINK
documentation})) into R}
}
\value{
A list containing three objects: an integer of chromosome of origin,
  an integer of SNP window center position, and a matrix of alternative
  allele counts for each genotype and SNP within the given window
}
\description{
Extract and format a window of SNPs to be tested in a SKAT kernel
}
\examples{
data("sample_genodata")

extract_window(this_position = 14460000,
  window_size = 3000,
  genodata = sample_genodata)

}
