% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtskat.R
\name{mtskat}
\alias{mtskat}
\title{Wrapper for SKAT sans resampling, with multi-threading over SNP windows}
\usage{
mtskat(
  this_phenotype,
  covariates,
  pre_allocated_SNP_windows,
  raw_file_path,
  n_thread,
  missing_cutoff = 0.15
)
}
\arguments{
\item{this_phenotype}{Phenotype data that has been read into R from the
standard PLINK phenotype file format
(\href{https://www.cog-genomics.org/plink2/formats}{PLINK documentation})}

\item{covariates}{Covariates data that has been read into R from a
comma-delimited text file with a header and one row for each genotype, in
the same order as the phenotype file.}

\item{pre_allocated_SNP_windows}{A list of SNP windows along with their
positions and scaffolds, prepared using \code{\link{pre_allocate}}}

\item{raw_file_path}{complete file path to SNP data in `.traw` format (see
\href{https://www.cog-genomics.org/plink2/formats}{PLINK documentation})}

\item{n_thread}{An integer indicating the number of threads to be used for
multithreading}

\item{missing_cutoff}{A numeric threshold representing the minimum desired
missing rate; missing rate is defined for each SNP as the proportion
of genotypes missing data for the given SNP. Imputation to mean is
performed , either by `pre_allocate` or `SKAT` itself,
for all remaining missing values}
}
\value{
A dataframe with columns for scaffold ID, SNP window position,
  p-values from the model used in SKAT without resampling, and an additional
  column for empirical p-values (all NA at this stage since resampling is not
  run in `mtskat`)
}
\description{
Wrapper for SKAT sans resampling, with multi-threading over SNP windows
}
\examples{
\dontrun{
data("sample_phenotype")
data("sample_covariates")
data("sample_pre_allocated_SNP_windows")

mtskat(
  this_phenotype = sample_phenotype,
  covariates = sample_covariates,
  raw_file_path = sample_pre_allocated_SNP_windows[[1]][[3]],
  window_size = 3000,
  window_shift = 1000,
  pre_allocated_SNP_windows = sampel_pre_allocated_SNP_windows,
  n_core = 2)
}
}
