% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_p_null.R
\name{tally_p_null}
\alias{tally_p_null}
\title{Multithreaded calculation of null p-values to enable calculation of empirical
p-values}
\usage{
tally_p_null(p_table)
}
\arguments{
\item{p_table}{a dataframe or matrix with a row for each SNP window and a
column for each p-value resulting from a permuted null model}
}
\value{

}
\description{
These functions are used to calculate empirical p-values given initial
p-values and null p-values from `SKAT`.
}
\details{
Prior to completion of jobs by cores in `MTMC-SKAT`, the number of p-values
above or below the in initial p-value are calculated using `tally_p_null`.
This reduces the load of data to be returned. Empirical p-values are then
calculated from all returned data using `p_empirical_from_tally`.
}
\examples{
\dontrun{sample_p_null_tallies <- tally_p_null(sample_p_table)}

}
